Move the DebugLogger to atlasdb-commons
