+++ atlasdb-commons/src/main/java/com/palantir/util/DebugLogger.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.palantir.timestamp;
+package com.palantir.util;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
