+++ timestamp-impl/build.gradle
@@ -8,9 +8,9 @@ dependencies {
   compile group: 'org.hamcrest', name: 'hamcrest-core'
   compile group: 'org.hamcrest', name: 'hamcrest-library'
 
+  testCompile project(":atlasdb-tests-shared")
   testCompile group: 'junit', name: 'junit'
   testCompile group: 'org.jmock', name: 'jmock', version: libVersions.jmock
   testCompile group: 'com.jayway.awaitility', name: 'awaitility'
   testCompile group: 'org.mockito', name: 'mockito-core'
-  testCompile group: 'org.assertj', name: 'assertj-core'
 }
