+++ timestamp-impl/src/main/java/com/palantir/timestamp/PersistentTimestampService.java
@@ -21,6 +21,7 @@ import javax.annotation.concurrent.ThreadSafe;
 
 import com.google.common.base.Preconditions;
 import com.palantir.common.concurrent.PTExecutors;
+import com.palantir.util.DebugLogger;
 
 @ThreadSafe
 public class PersistentTimestampService implements TimestampService, TimestampMigrationService {
