+++ timestamp-impl/src/main/java/com/palantir/timestamp/PersistentUpperLimit.java
@@ -24,6 +24,7 @@ import javax.annotation.concurrent.GuardedBy;
 import com.palantir.common.time.Clock;
 import com.palantir.common.time.SystemClock;
 import com.palantir.exception.PalantirInterruptedException;
+import com.palantir.util.DebugLogger;
 
 public class PersistentUpperLimit {
     private final TimestampBoundStore tbs;
